Move base material seeding and update seeded inventories
